리펙토링
- GameData 메소드로 카운트 계산 메소드 구현
- 정답을 맞춘 경우 다시 한번 정답을 제출했을 때 남은 횟수를 0으로 만들어
  게임 종료 메세지를 리턴
